refactor(wallet-sdk): move User types + predicates to SDK (Phase 3.1)
Moves User, FullUser, GuestUser, UserProfile, UpdateUser, and the three
predicate functions into the SDK under packages/wallet-sdk/src/user/user.ts.

Replaces the structural UserDefaultAccounts workaround in AccountService
with the real User type now that it lives in the SDK.

Web user.ts becomes a transitional re-export shim.
