+++ apps/web-wallet/app/features/user/user.ts
@@ -1,42 +1,13 @@
-import type { Currency } from '~/lib/money';
-
-type CommonUserData = {
-  id: string;
-  username: string;
-  emailVerified: boolean;
-  createdAt: string;
-  updatedAt: string;
-  defaultBtcAccountId: string;
-  defaultUsdAccountId: string | null;
-  defaultCurrency: Currency;
-  cashuLockingXpub: string;
-  encryptionPublicKey: string;
-  sparkIdentityPublicKey: string;
-  termsAcceptedAt: string | null;
-  giftCardMintTermsAcceptedAt: string | null;
-};
-
-export type FullUser = CommonUserData & {
-  email: string;
-  isGuest: false;
-};
-
-export type GuestUser = CommonUserData & {
-  isGuest: true;
-};
-
-export type User = FullUser | GuestUser;
-
-export type UserProfile = Pick<User, 'id' | 'username'>;
-
-export function shouldVerifyEmail(user: User): user is FullUser {
-  return !user.isGuest && !user.emailVerified;
-}
-
-export function shouldAcceptTerms(user: User): boolean {
-  return user.termsAcceptedAt === null;
-}
-
-export function shouldAcceptGiftCardMintTerms(user: User): boolean {
-  return user.giftCardMintTermsAcceptedAt === null;
-}
+// Transitional re-export — moved to @agicash/wallet-sdk; removed in the import-cleanup PR.
+export type {
+  FullUser,
+  GuestUser,
+  User,
+  UserProfile,
+  UpdateUser,
+} from '@agicash/wallet-sdk';
+export {
+  shouldVerifyEmail,
+  shouldAcceptTerms,
+  shouldAcceptGiftCardMintTerms,
+} from '@agicash/wallet-sdk';
