+++ packages/wallet-sdk/src/user/user.ts
@@ -1,0 +1,51 @@
+import type { Currency } from '@agicash/utils/money';
+
+type CommonUserData = {
+  id: string;
+  username: string;
+  emailVerified: boolean;
+  createdAt: string;
+  updatedAt: string;
+  defaultBtcAccountId: string;
+  defaultUsdAccountId: string | null;
+  defaultCurrency: Currency;
+  cashuLockingXpub: string;
+  encryptionPublicKey: string;
+  sparkIdentityPublicKey: string;
+  termsAcceptedAt: string | null;
+  giftCardMintTermsAcceptedAt: string | null;
+};
+
+export type FullUser = CommonUserData & {
+  email: string;
+  isGuest: false;
+};
+
+export type GuestUser = CommonUserData & {
+  isGuest: true;
+};
+
+export type User = FullUser | GuestUser;
+
+export type UserProfile = Pick<User, 'id' | 'username'>;
+
+export type UpdateUser = {
+  defaultBtcAccountId?: string;
+  defaultUsdAccountId?: string | null;
+  defaultCurrency?: Currency;
+  username?: string;
+  termsAcceptedAt?: string;
+  giftCardMintTermsAcceptedAt?: string;
+};
+
+export function shouldVerifyEmail(user: User): user is FullUser {
+  return !user.isGuest && !user.emailVerified;
+}
+
+export function shouldAcceptTerms(user: User): boolean {
+  return user.termsAcceptedAt === null;
+}
+
+export function shouldAcceptGiftCardMintTerms(user: User): boolean {
+  return user.giftCardMintTermsAcceptedAt === null;
+}
